+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## [Unreleased]
 #### Fixed
+- Changed Activity type used as parameter in 'PushNotification.accept' [SDKS-1968]
 - Deserializing an object with whitelist to prevent deserialization of untrusted data.  [SDKS-1818]
+
 
 ## [3.3.3]
 #### Fixed