+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Updated `authenticator` module and sample app to handle the new `POST_NOTIFICATIONS` permission. [SDKS-2033]
 - DefaultTokenManager does not cache AccessToken in memory upon retrieval from Shared Preferences.  [SDKS-2066]
 - Align forgerock_logout_endpoint configuration name with iOS [SDKS-2085]
+- Allow leading slash on custom endpoint path [SDKS-2074]
 
 ## [3.3.3]
 #### Fixed